Reduce tracking-from-scratch to creating an empty tree in the Tracks List tab
